requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
contourpy=1.1.0=pypi_0
cupy-cuda12x=12.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py311h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fastrlock=0.8.1=pypi_0
fonttools=4.40.0=pypi_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.14.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h38be061_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvmlite=0.40.1=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numba=0.57.1=pypi_0
numpy=1.24.4=pypi_0
openssl=3.0.9=h7f8727e_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.0=pypi_0
pip=23.1.2=py311h06a4308_0
platformdirs=3.8.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pyparsing=3.1.0=pypi_0
python=3.11.3=h955ad1f_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytz=2023.3=pypi_0
pyzmq=25.1.0=py311h6a678d5_0
readline=8.2=h5eee18b_0
setuptools=67.8.0=py311h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=py311h5eee18b_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.0=hd8ed1ab_0
typing_extensions=4.7.0=pyha770c72_0
tzdata=2023.3=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
